+++ package.json
@@ -36,7 +36,6 @@
     "browser-stdout": "^1.3.0",
     "es6-shim": "^0.35.1",
     "font-awesome": "^4.6.3",
-    "handlebars": "^4.0.5",
     "jquery": "^2.2.2",
     "json-loader": "^0.5.4",
     "ng2-select2": "0.3.1",
