+++ src/components/keymap/key-map.component.html
@@ -1,0 +1,35 @@
+<div class="row">
+    <h1 class="col-xs-12 pane-title">
+        <i class="fa fa-keyboard-o"></i>
+        <span class="keymap__name pane-title__name" contenteditable="true">QWERTY</span> keymap
+        (<span class="keymap__abbrev pane-title__abbrev" contenteditable="true">QTY</span>)
+        <i class="fa fa-star-o fa-star keymap__is-default"></i>
+        <i class="glyphicon glyphicon-trash keymap__remove pull-right" title=""
+           data-toggle="tooltip" data-placement="left" data-original-title="Remove keymap"></i>
+    </h1>
+</div>
+<div class="row uhk--wrapper">
+    <div class="col-xs-12 text-center">
+        <span class="uhk__layer-switcher--wrapper" data-title="Layers: ">
+            <button #baseButton type="button" class="btn btn-default btn-lg btn-primary" (click)="selectLayer(0)">
+                Base
+            </button>
+            <button #modButton type="button" class="btn btn-default btn-lg" (click)="selectLayer(1)">
+                Mod
+            </button>
+            <button #fnButton type="button" class="btn btn-default btn-lg" (click)="selectLayer(2)">
+                Fn
+            </button>
+            <button #mouseButton type="button" class="btn btn-default btn-lg" (click)="selectLayer(3)">
+                Mouse
+            </button>
+        </span>
+    </div>
+    <div class="keyboard-slider">
+        <svg-keyboard-popover *ngFor="let layer of layers.elements"
+                              [moduleConfig]="layer.modules.elements"
+                              (animationend)="onKeyboardAnimationEnd($event)"
+                              hidden>
+        </svg-keyboard-popover>
+    </div>
+</div>
