+++ src/components/legacy/legacy-loader.component.scss
@@ -1,0 +1,10 @@
+:host {
+    display: none;
+    height: 100%;
+    width: 100%;
+}
+
+iframe{
+    height: 100vh;
+    width: 100%;
+}
