+++ src/components/sidemenu/side-menu.component.html
@@ -1,0 +1,25 @@
+<ul class="menu--top legacy">
+    <li class="sidebar__level-1--item">
+        <div class="sidebar__level-1">
+            <i class="fa fa-exclamation-triangle "></i> Legacy
+            <i class="fa fa-chevron-up pull-right" style=""></i>
+        </div>
+        <ul>
+            <li class="sidebar__level-2--item" data-name="Keymap" data-content="key-map" data-abbrev="QTY">
+                <div class="sidebar__level-2">
+                    Keymap Angular
+                </div>
+            </li>
+            <li class="sidebar__level-2--item" data-name="Keymap" data-content=".keymap-legacy--edit" data-abbrev="QTY">
+                <div class="sidebar__level-2">
+                    Keymap HTML
+                </div>
+            </li>
+            <li class="sidebar__level-2--item " data-name="Macro" data-content=".macro-legacy--edit" data-abbrev="DVR">
+                <div class="sidebar__level-2">
+                    Macro
+                </div>
+            </li>
+        </ul>
+    </li>
+</ul>
