+++ src/components/sidemenu/side-menu.component.scss
@@ -1,0 +1,8 @@
+:host {
+    background-color: #f5f5f5;
+    border-right: 1px solid #ccc;
+    position: fixed;
+    overflow-y: auto;
+    width: 250px;
+    height: 100%;
+}
