+++ src/components/sidemenu/side-menu.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'side-menu',
+    templateUrl: 'src/components/sidemenu/side-menu.component.html',
+    styles: [require('./side-menu.component.scss')]
+})
+export class SideMenuComponent {
+    constructor() {
+    }
+}
